Resolvendo problemas de duplicacao

+++ mtds_U2/sistemas_lineares/utils.py
@@ -5,9 +5,9 @@ def resolve_inferior(A, b):
     assumindo que A é uma matriz triangular inferior.
     '''
     N = len(A)
-    x = np.zeros((N, 1))
+    x = np.zeros((N, 1), dtype=float)
     for i in range(0, N):
-        acc =  0
+        acc =  0.0
         for j in range(0, i):
             acc += A[i][j] * x[j][0]
         
@@ -22,9 +22,9 @@ def resolve_superior(A, b):
     assumindo que A é uma matriz triangular superior.
     '''
     N = len(A)
-    x = np.zeros((N, 1))
+    x = np.zeros((N, 1), dtype=float)
     for i in range(N-1, -1, -1): # de N-1 até 0
-        acc = 0
+        acc = 0.0
         for j in range(i+1, N):
             acc += A[i][j] * x[j][0]
